requirements.txt
aiofiles==24.1.0
alembic==1.16.5
anyio==4.11.0
asyncpg==0.30.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.8.3
click==8.3.0
cryptg==0.5.1
Flask==3.1.2
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
priority==2.0.0
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.1
Quart==0.20.0
Quart-Session==3.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
Telethon==1.41.2
typing_extensions==4.15.0
uvicorn==0.37.0
Werkzeug==3.1.3
wsproto==1.2.0
